Add HttpRequestProcessor
renamed HttpRequestForTasks to TasksHttpRequest
renamed HttpRequestMethod to TasksHttpRequestMethod
