+++ app/src/main/java/com/codesoom/assignment/models/HttpRequestProcessor.java
@@ -1,0 +1,7 @@
+package com.codesoom.assignment.models;
+
+import com.codesoom.assignment.service.TaskService;
+
+public interface HttpRequestProcessor {
+    HttpResponse processTasks(HttpRequest httpRequest, TaskService taskService);
+}
